Module Core

Types

VPL.Core.Graph — Type

Graph(axiom; rules = nothing, vars = nothing)

Creates a dynamic graph defined by the initial node or nodes (axiom), one or more rules (rules), and an object with graph-level variables (vars). Rules and graph-level variables are optional and must be assigned by keyword (see example below). Rules must be an Rule or tuple of Rule objects. The axiom may be a single object inheriting from Node or a subgraph generated with the graph construction DSL. A copy of the axiom and rules is always made when constructing the graph, but if object containing graph-level variables is not mutable, the user must manually copy (with copy or deepcopy) or else changes within the graph will affect the original object (and other graphs creates from the same object).

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
no_rules_graph = Graph(axiom)
rule = Rule(A, rhs = x -> A() + B())
rules_graph = Graph(axiom, rules = rule)

VPL.Core.Rule — Type

Rule(nodetype; lhs = x -> true, rhs = x -> nothing, captures = false)

Create a replacement rule for nodes of type nodetype with function-like objects for the left-hand side (lhs) and right-hand side (rhs). If the rule captures nodes in the context of the replacement node this must be indicated by the argument captures.

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
rule = Rule(A, rhs = x -> A() + B())
rules_graph = Graph(axiom, rules = rule)
rewrite!(rules_graph)

VPL.Core.Query — Type

Query(nodetype::DataType, query = x -> true)

Create a query that matches nodes of type nodetype and the conditions specified in the argument query (must be a function that returns true). It returns an object of type Query that can be applied to a graph with the function apply.

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
graph = Graph(axiom)
query = Query(A)
apply(graph, query)

VPL.Core.Node — Type

Node

Abstract type from which every node in a graph should inherit. This allows using the graph construction DSL

Example

struct bar <: Node
  x::Int
end
b1 = bar(1)
b2 = bar(2)
b1 + b2

VPL.Core.Context — Type

Context

Data structure than links a GraphNode to a Graph. Functions data and vars give access to the data stored in the node and graph, respectively. Several methods are also available to test relationships among nodes in the graph and to extract these related nodes (see User Manual for details).

Uses do not build Context objects directly but they are provided by VPL as inputs to the user-defined functions inside rules and queries.

Applying rules and queries

VPL.Core.apply — Method

apply(g::Graph, query::Query)

Return an array with all the nodes in the graph that match the query supplied by the user.

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
graph = Graph(axiom)
query = Query(A)
apply(graph, query)

VPL.Core.rewrite! — Method

rewrite!(g::Graph)

Apply the graph-rewriting rules stored in the graph. This function will match the left-hand sides of the rules against the graph and then replace and/or prune the graph at every location where the left-hand sides matched by the result of executing the right hand side of each rule. The modification is performed in-place, so this function returns nothing.

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
rule = Rule(A, rhs = x -> A() + B())
rules_graph = Graph(axiom, rules = rule)
rewrite!(rules_graph)

Extracting information

VPL.Core.vars — Method

vars(g::Graph)

Returns the object storing the graph-level variables

Example

struct A <: Node end
axiom = A()
graph = Graph(axiom, vars = 2)
vars(graph)

VPL.Core.rules — Method

rules(g::Graph)

Returns a tuple with all the graph-rewriting rules stored in the graph

Example

struct A <: Node end
struct B <: Node end
axiom = A() + B()
rule = Rule(A, rhs = x -> A() + B())
rules_graph = Graph(axiom, rules = rule)
rules(rules_graph)

VPL.Core.vars — Method

vars(c::Context)

Returns the object storing the graph-level variables in the graph associated to a Context object. This needs to be used inside rules and queries.

VPL.Core.data — Method

data(c::Context)

Returns the data stored in the node associated to a Context object. This needs to be used inside rules and queries.

Graph traversal

VPL.Core.hasParent — Method

hasParent(c::Context)

Check if the node passed as argument has a parent and return true or false.

VPL.Core.isRoot — Method

isRoot(c::Context)

Check if the node passed as argument is the root of the graph (i.e. has no parent) and return true or false.

VPL.Core.hasAncestor — Function

hasAncestor(c::Context, condition, maxlevel)

Check if the node passed as argument has an ancestor that matches the optional condition and and return true or false and the number of steps taken. The argument maxlevel is optional and limits the number of steps that the algorithm will move through the graph (by default there is no limitation). The default condition returns true for any ancestor and it takes an object of type Context.

Base.parent — Method

parent(c::Context, nsteps::Int)

Returns a Context object associated to the parent of the node passed as first argument (nsteps = 1, the default) or an ancestor that is nsteps away from the node passed as first argument.

VPL.Core.ancestor — Function

ancestor(c::Context, condition, maxlevel)

Returns a Context object associated to the first ancestor of the node given as argument that matches the optional condition. The argument maxlevel is optional and limits the number of steps that the algorithm will move through the graph (by default there is no limitation). The matched node is returned as a Context object. The default condition returns true for any ancestor and it takes an object of type Context.

VPL.Core.hasChildren — Method

hasChildren(c::Context)

Check if the node passed as argument has at least one child and return true or false.

VPL.Core.isLeaf — Method

isLeaf(c::Context)

Check if the node passed as argument is a leaf in the graph (i.e. has no children) and return true or false.

VPL.Core.hasDescendent — Function

hasDescendent(c::Context, condition, maxlevel)

Check if the node passed as argument has a descendent that matches the optional condition and and return true or false. The argument maxlevel is optional and limits the number of steps that the algorithm will move through the graph (by default there is no limitation). The default condition returns true for any descendent and it takes an object of type Context.

VPL.Core.children — Method

children(c::Context)

Returns a tuple of Context objects with all the children of thew node given as argument.

VPL.Core.descendent — Function

descendent(c::Context, condition, maxlevel)

Returns a Context object associated to the first descendent of the node given as argument that matches the optional condition. The argument maxlevel is optional and limits the number of steps that the algorithm will move through the graph (by default there is no limitation). The matched node is returned as a Context object. The default condition returns true for any descendent and it takes an object of type Context.

VPL.Core.traverse — Method

traverse(g::Graph, f)

Iterates over all the nodes in the graph (in no particular order) and execute for each node the function f taking as input the data stored in the node.

VPL.Core.traverseDFS — Method

traverseDFS(g::Graph, f)

Iterates over all the nodes in the graph (depth-first order, starting at the root of the graph) and execute for each node the function f taking as input the data stored in the node.

VPL.Core.traverseBFS — Method

traverseBFS(g::Graph, f)

Iterates over all the nodes in the graph (breadth-first order, starting at the root of the graph) and execute for each node the function f taking as input the data stored in the node.

Graph visualization

VPL.Core.draw — Method

draw(g::Graph; name::String = "VPL Graph")

Draw a network representation of the graph. The drawing is performed on a Blink window using the vis.js Javascript library. The function returns the handler to the Blink window. Access to Internet is required as the libraries are loaded from the cdns server.

VPL.Core.get_id — Method

get_id(key, data)

Method that defines the label associated to the node data when drawing the graph. The argument key is the unique identifier of the node in the graph. The default method prints the type of data stored in the node and the unique identifier of the node in parenthesis.